Modules for preparing xyz file for Prismatic
It contains functions to orient and orthogonalise the cell from any
cif file (or any file supported by the ASE package) and convert it
to the xyz file format suitable for Prismatic. It also contains
functions to run simulation through pyprismatic. A demo notebook
using a triclinic cyrstal of AlPO4 is included.